Changed DeepLUsageData fields format

+++ src/main/java/com/subtitlescorrector/core/domain/deepl/DeepLUsageData.java
@@ -5,20 +5,20 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 public class DeepLUsageData {
 	
 	@JsonProperty("character_count")
-	Integer characterCount;
+	Long characterCount;
 	@JsonProperty("character_limit")
-	Integer characterLimit;
+	Long characterLimit;
 	
-	public Integer getCharacterCount() {
+	public Long getCharacterCount() {
 		return characterCount;
 	}
-	public void setCharacterCount(Integer characterCount) {
+	public void setCharacterCount(Long characterCount) {
 		this.characterCount = characterCount;
 	}
-	public Integer getCharacterLimit() {
+	public Long getCharacterLimit() {
 		return characterLimit;
 	}
-	public void setCharacterLimit(Integer characterLimit) {
+	public void setCharacterLimit(Long characterLimit) {
 		this.characterLimit = characterLimit;
 	}
